refractor: move index mapping to common components
