+++ public/components/IndexMapping/IndexMapping.scss
