+++ public/components/IndexMapping/IndexMapping.test.tsx
@@ -7,7 +7,7 @@ import React, { forwardRef, Ref, useRef, useState } from "react";
 import { render, fireEvent, waitFor, act } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 import IndexMapping, { IIndexMappingsRef, IndexMappingProps, transformObjectToArray } from "./IndexMapping";
-import { MappingsProperties } from "../../../../../models/interfaces";
+import { MappingsProperties } from "../../../models/interfaces";
 import { renderHook } from "@testing-library/react-hooks";
 
 const IndexMappingOnChangeWrapper = forwardRef((props: Partial<IndexMappingProps>, ref: Ref<IIndexMappingsRef>) => {
