+++ public/components/IndexMapping/index.ts
